nxlogger: adjust build for new nexus interface.
note: we force a dependency on nexus in order to pick up the
      generated header which would otherwise fail the build.

refs #SWSTB-1530

+++ nxif/nxlogger/nxlogger.cpp
@@ -130,7 +130,7 @@ static BERR_Code get_driver_log_message(int device_fd, PROXY_NEXUS_Log_Instance 
 
     dequeue.instance    = *pProxyInstance;
     dequeue.buffer_size = bufLen-1;
-    dequeue.buffer      = pBuf;
+    dequeue.buffer      = (uint64_t)(intptr_t)pBuf;
     dequeue.timeout     = 0;
     urc = ioctl(device_fd, IOCTL_PROXY_NEXUS_Log_Dequeue, &dequeue);
     if(urc!=0) {
